ai-service/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
prisma==0.12.0

# Message Queue
pika==1.3.2
aio-pika==9.3.1

# Storage
minio==7.2.3
boto3==1.34.14

# AI & ML
httpx==0.26.0
pillow==10.2.0
numpy==1.26.3

# LLM
google-generativeai==0.3.2
openai==1.8.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
redis==5.0.1
structlog==24.1.0
